temp

+++ src/hotspot/cpu/x86/methodHandles_x86.cpp
@@ -467,7 +467,7 @@ void MethodHandles::generate_method_handle_dispatch(MacroAssembler* _masm,
     }
 
     default:
-      fatal("unexpected intrinsic %d: %s", iid, vmIntrinsics::name_at(iid));
+      fatal("unexpected intrinsic %d: %s", (int)iid, vmIntrinsics::name_at(iid));
       break;
     }
