+++ src/hotspot/cpu/x86/sharedRuntime_x86_64.cpp
@@ -1849,7 +1849,7 @@ static void gen_special_dispatch(MacroAssembler* masm,
   } else if (iid == vmIntrinsics::_invokeBasic) {
     has_receiver = true;
   } else {
-    fatal("unexpected intrinsic id %d", iid);
+    fatal("unexpected intrinsic id %d", (int)iid);
   }
 
   if (member_reg != noreg) {
